ReactRouter導入
・What
URLを変更すれば該当するページに遷移できるようにした。
pages/indexはページの切り替えだけに専念するページにする
⇨レンダリングするページはtemplates配下のディレクトリに移動
