+++ frontend/src/components/templates/dishes.jsx
@@ -1,0 +1,10 @@
+// pagesのディレクトリに実際にrenderするファイルをおく
+import React , { Component, useState } from 'react';
+
+
+export const Dishes = () => {
+  return (
+      <div>aaa</div>
+  );
+};
+
