+++ frontend/src/components/templates/style.scss
